[deckhouse] Move app config package into deckhouse-controller (#21695)
