+++ deckhouse-controller/internal/app/globals.go
