+++ deckhouse-controller/internal/app/paths.go
