+++ deckhouse-controller/internal/app/resources.go
